[Refactor/#4] 폴더명 변경
