+++ app/src/main/java/org/android/go/sopt/present/SearchFragment.kt
@@ -1,28 +1,26 @@
-package org.android.go.sopt.home
+package org.android.go.sopt.present
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import org.android.go.sopt.databinding.FragmentGalleryBinding
-import org.android.go.sopt.databinding.FragmentHomeBinding
+import org.android.go.sopt.databinding.FragmentSearchBinding
 
-class GalleryFragment: Fragment() {
-    private var _binding: FragmentGalleryBinding? = null
-    private val binding: FragmentGalleryBinding
+class SearchFragment: Fragment() {
+    private var _binding: FragmentSearchBinding? = null
+    private val binding: FragmentSearchBinding
         get() = requireNotNull(_binding) { "앗 ! _binding이 null이다 !" }
 
-    override fun onCreateView( // 뷰를 만든다 << 이때 초기화하면 좋음
+    override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View { // 이제 반환하는 View가 Null일 수 없기 때문에, ?를 지워주셔도 됩니다.
-        _binding = FragmentGalleryBinding.inflate(inflater, container, false)
+        _binding = FragmentSearchBinding.inflate(inflater, container, false)
         return binding.root
     }
 
-    // 뷰가 만들어졌다.
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         // 대부분의 로직은 여기에 구현합니다.
